+++ Projects/Feature/Sources/View/Edit/ChangeTeamInfoView.swift
@@ -1,0 +1,19 @@
+//
+//  ChangeTeamInfoView.swift
+//  Feature
+//
+//  Created by 한지석 on 11/21/23.
+//  Copyright © 2023 com.pivoters. All rights reserved.
+//
+
+import SwiftUI
+
+struct ChangeTeamInfoView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    ChangeTeamInfoView()
+}
